pyproject.toml
[project]
name = "QMDown"
dynamic = ["version"]
description = "QQ音乐解析/下载工具"
authors = [
    {name = "Luren", email = "68656403+luren-dc@users.noreply.github.com"},
]
dependencies = [
    "typer>=0.13.0",
    "qqmusic-api-python>=0.3.1",
    "httpx>=0.27.2",
    "pydantic>=2.10.4",
    "pillow>=11.0.0",
    "pyzbar>=0.1.9",
    "qrcode>=8.0",
    "tenacity>=9.0.0",
    "mutagen>=1.47.0",
]
requires-python = "<3.13,>=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
QMDown = "QMDown.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.hatch.version]
path = "QMDown/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["QMDown"]

[tool.hatch.build.targets.sdist]
include = ["/QMDown", "/tests", "LISENCE", "README.md"]

[tool.commitizen]
name = "cz_gitmoji"

[tool.basedpyright]
venvPath = "."
venv = ".venv"
include = ["QMDown"]
exclude = ["tests"]
pythonVersion = "3.10"
typeCheckingMode = "standard"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
extend-select = ["Q", "RUF", "C90", "UP", "I", "ASYNC", "C4", "FURB", "R"]
pydocstyle = { convention = "google" }

[dependency-groups]
dev = []
